bt-670: * fix navigatiom
