+++ frontend/src/bundles/common/pages/home.tsx
@@ -4,7 +4,7 @@ import { type RootReducer } from '~/framework/store/store.package.js';
 import { Navigate } from '../components/components.js';
 import { AppRoute } from '../enums/app-route.enum.js';
 import { configureString } from '../helpers/helpers.js';
-import { useAppSelector } from '../hooks/hooks.js';
+import { useAppSelector, useEffect, useNavigate } from '../hooks/hooks.js';
 import { UserRole } from '../types/types.js';
 
 const Home: React.FC = () => {
@@ -16,14 +16,18 @@ const Home: React.FC = () => {
             ? state.talentOnBoarding
             : state.employerOnBoarding,
     );
+    const navigate = useNavigate();
+    useEffect(() => {
+        if (isApproved) {
+            navigate(AppRoute.CHATS);
+        }
+    }, [isApproved, navigate]);
+
     switch (role) {
         case UserRole.ADMIN: {
             return <Navigate to={AppRoute.ADMIN_VERIFICATIONS_PANEL} />;
         }
         case UserRole.TALENT: {
-            if (isApproved) {
-                return <Navigate to={AppRoute.CHATS} />;
-            }
             return (
                 <Navigate
                     to={configureString(AppRoute.TALENT_STEP, {
@@ -33,13 +37,10 @@ const Home: React.FC = () => {
             );
         }
         case UserRole.EMPLOYER: {
-            if (isApproved) {
-                return <Navigate to={AppRoute.CHATS} />;
-            }
-            return <Navigate to={AppRoute.CANDIDATES} />;
+            return <Navigate to={AppRoute.EMPLOYER_ONBOARDING} />;
         }
         default: {
-            return <Navigate to={AppRoute.NOT_FOUND} />;
+            return <Navigate to={AppRoute.SIGN_IN} />;
         }
     }
 };
